Refactor Gamepad View
