+++ client/src/components/views/GamepadView/exports.ts
@@ -1,0 +1,14 @@
+// Main component
+export { default } from './index';
+
+// Subcomponents
+export { GamepadButton } from './GamepadButton';
+export { GamepadStick } from './GamepadStick';
+export { GamepadControls } from './GamepadControls';
+export { GamepadHeader } from './GamepadHeader';
+export { GamepadToolbar } from './GamepadToolbar';
+export { GamepadSettings } from './GamepadSettings';
+
+// Hooks
+export { useGamepadState } from './hooks/useGamepadState';
+export { useKeyboardControls } from './hooks/useKeyboardControls';
